Add files via upload
Main files you need to look at. Components is the file we are using to interact with gapi, the google calendar api tool used to import and export events as APIs. Index is the default file generated when nextJS app is made. We are going to use it later with the FullCalendar packages to make a UI for choosing the time slot.